+++ src/module.ts
@@ -1,11 +1,8 @@
 import CogniteDatasource from './datasource';
-import { CogniteAnnotationsQueryCtrl, CogniteQueryCtrl } from './query_ctrl';
+import {CogniteQueryCtrl} from './query_ctrl';
 import {CogniteConfigCtrl} from './config_ctrl';
+import {CogniteAnnotationsQueryCtrl} from './annotation_ctrl';
 import {CogniteVariableQueryCtrl} from './variable_query_ctrl';
-
-// class ChangeMyNameAnnotationsQueryCtrl {
-//   static templateUrl = 'partials/annotations.editor.html';
-// }
 
 export {
   CogniteDatasource as Datasource,
@@ -13,5 +10,4 @@ export {
   CogniteConfigCtrl as ConfigCtrl,
   CogniteAnnotationsQueryCtrl as AnnotationsQueryCtrl,
   CogniteVariableQueryCtrl as VariableQueryEditor,
-  //ChangeMyNameAnnotationsQueryCtrl as AnnotationsQueryCtrl,
 };
